typo fix
I believe that the sentence is speaking about "the importer", aka
the person doing the importing, not "the import", possibly aka the
import statement above.

I guess "import" might be correct if one believe that the import
statement in a source file a) had a gender and b) was male.